pyproject.toml
[project]
name = "honeypot-mcp"
version = "0.1.0"
description = "AI Red Team + Content Scorer + Research Agent — MCP server powered by Gemini Flash"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
